Adds common value conversions and mechanisms for type mapping to pick one
Part of #242

This change means that providers do not need to add mappings involving conversions. Instead providers will only add mappings for specific types that they handle "natively" and if the mapping request for a different type is made, then the type mapper will find a conversion to a type that is supported and automatically use it.

A converter is also automatically found if the model type and the store type specified in the model do not match.

Ultimately, any "common" type can always be converted to string or bytes, so as long as a provider can map at least one of these, then any common type can always be mapped. There are two tests for this that create type mappers for SQL Server that support only strings or only bytes. The database is then created and the tests are run against only nvarchar or only varbinary columns, even though the model contains all common types.